admin delete authorization

+++ app/routes/users+/$username_+/notes.$noteId.tsx
@@ -21,7 +21,7 @@ import { ErrorList } from '#app/components/forms.tsx'
 import { Button } from '#app/components/ui/button.tsx'
 import { Icon } from '#app/components/ui/icon.tsx'
 import { StatusButton } from '#app/components/ui/status-button.tsx'
-import { requireUser } from '#app/utils/auth.server.ts'
+import { getUserId, requireUser } from '#app/utils/auth.server.ts'
 import { validateCSRF } from '#app/utils/csrf.server.ts'
 import { prisma } from '#app/utils/db.server.ts'
 import {
@@ -33,9 +33,10 @@ import { redirectWithToast } from '#app/utils/toast.server.ts'
 import { useOptionalUser } from '#app/utils/user.ts'
 import { type loader as notesLoader } from './notes.tsx'
 
-export async function loader({ params }: LoaderFunctionArgs) {
+export async function loader({ request, params }: LoaderFunctionArgs) {
 	// 🐨 get the user id from the request (remember, unauthenticated users should
 	// be able to see these notes too, so use getUserId instead of requireUserId).
+	const userId = await getUserId(request)
 	const note = await prisma.note.findUnique({
 		where: { id: params.noteId },
 		select: {
@@ -66,11 +67,29 @@ export async function loader({ params }: LoaderFunctionArgs) {
 	// 📜 https://www.prisma.io/docs/reference/api-reference/prisma-client-reference#some
 	// 💰 you have the note.ownerId, so you can use that to decide whether you
 	// should be looking for "access: 'own'" or "access: 'any'".
+	const permission = userId
+		? await prisma.permission.findFirst({
+				select: { id: true },
+				where: {
+					roles: {
+						some: {
+							users: {
+								some: { id: userId },
+							},
+						},
+					},
+					entity: 'note',
+					action: 'delete',
+					access: note.ownerId === userId ? 'own' : 'any',
+				},
+			})
+		: null
 
 	return json({
 		note,
 		timeAgo,
 		// 🐨 set canDelete to whether there is a matching permission
+		canDelete: Boolean(permission),
 	})
 }
 
@@ -82,9 +101,9 @@ const DeleteFormSchema = z.object({
 export async function action({ request, params }: ActionFunctionArgs) {
 	const user = await requireUser(request)
 	// 💣 since admins can delete notes, remove this
-	invariantResponse(user.username === params.username, 'Not authorized', {
+	/* invariantResponse(user.username === params.username, 'Not authorized', {
 		status: 403,
-	})
+	}) */
 	const formData = await request.formData()
 	await validateCSRF(formData, request.headers)
 	const submission = parse(formData, {
@@ -101,16 +120,34 @@ export async function action({ request, params }: ActionFunctionArgs) {
 
 	const note = await prisma.note.findFirst({
 		// 🐨 now we need the ownerId as well so we can determine whether the current user is the owner
-		select: { id: true, owner: { select: { username: true } } },
+		select: { id: true, ownerId: true, owner: { select: { username: true } } },
 		// 💣 since admins can delete notes, we can't filter by userId anymore, remove that
-		where: { id: noteId, ownerId: user.id },
+		where: { id: noteId },
 	})
 	invariantResponse(note, 'Not found', { status: 404 })
 
 	// 🐨 do an identical permission query to the one above.
+	const permission = await prisma.permission.findFirst({
+		select: { id: true },
+		where: {
+			roles: { some: { users: { some: { id: user.id } } } },
+			entity: 'note',
+			action: 'delete',
+			access: note.ownerId === user.id ? 'own' : 'any',
+		},
+	})
 
 	// 🐨 if there is no permission, then throw a response with a 403 status code
 	// which you can handle in the error boundary below
+	if (!permission) {
+		throw json(
+			{
+				error: 'Unauthorized',
+				message: `Unauthorized: requires permission delete:note`,
+			},
+			{ status: 403 },
+		)
+	}
 
 	await prisma.note.delete({ where: { id: note.id } })
 
@@ -126,7 +163,7 @@ export default function NoteRoute() {
 	const user = useOptionalUser()
 	const isOwner = user?.id === data.note.ownerId
 	// 🐨 get this value from the loader data
-	const canDelete = true
+	const canDelete = data.canDelete
 	const displayBar = canDelete || isOwner
 
 	return (
